Ticket added

+++ ticket_handlers.go
@@ -19,6 +19,17 @@ func validateTicketData(w http.ResponseWriter, t Ticket) bool {
 		http.Error(w, "Цена не может быть отрицательной", http.StatusBadRequest)
 		return false
 	}
+
+	if err := uuid.Validate(t.MovieShowID); err != nil {
+		http.Error(w, "Неверный формат ID сеанса", http.StatusBadRequest)
+		return false
+	}
+
+	if err := uuid.Validate(t.SeatID); err != nil {
+		http.Error(w, "Неверный формат ID места", http.StatusBadRequest)
+		return false
+	}
+
 	return true
 }
 
@@ -50,7 +61,7 @@ func CreateTicket(db *pgxpool.Pool) http.HandlerFunc {
 			return
 		}
 
-		_, err := db.Exec(context.Background(), "INSERT INTO tickets (id, movie_show_id, seat_id, ticket_status_id, price) VALUES ($1, $2, $3, $4, $5)",
+		_, err := db.Exec(context.Background(), "INSERT INTO tickets (id, movie_show_id, seat_id, ticket_status, price) VALUES ($1, $2, $3, $4, $5)",
 			t.ID, t.MovieShowID, t.SeatID, t.Status, t.Price)
 		if IsError(w, err) {
 			return
@@ -66,13 +77,16 @@ func CreateTicket(db *pgxpool.Pool) http.HandlerFunc {
 // @Param movie_show_id path string true "ID показа фильма"
 // @Success 200 {array} Ticket
 // @Failure 500 {object} ErrorResponse "Ошибка сервера"
-// @Router /seats/{movie_show_id} [get]
+// @Router /tickets/movie-show/{movie_show_id} [get]
 func GetTicketsByMovieShowID(db *pgxpool.Pool) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		movieShowID := r.PathValue("movie_show_id")
+		movieShowID, ok := ParseUUIDFromPath(w, r.PathValue("movie_show_id"))
+		if !ok {
+			return
+		}
 
 		rows, err := db.Query(context.Background(), `
-			SELECT t.id, t.movie_show_id, t.seat_id, t.ticket_status_id, t.price
+			SELECT t.id, t.movie_show_id, t.seat_id, t.ticket_status, t.price
 			FROM tickets t
 			WHERE t.movie_show_id = $1`, movieShowID)
 		if HandleDatabaseError(w, err, "билетами") {
@@ -107,9 +121,12 @@ func GetTicketsByMovieShowID(db *pgxpool.Pool) http.HandlerFunc {
 // @Router /tickets/{id} [get]
 func GetTicketByID(db *pgxpool.Pool) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		id := r.PathValue("id")
+		id, ok := ParseUUIDFromPath(w, r.PathValue("id"))
+		if !ok {
+			return
+		}
 		var t Ticket
-		err := db.QueryRow(context.Background(), "SELECT id, movie_show_id, seat_id, ticket_status_id, price FROM tickets WHERE id = $1", id).
+		err := db.QueryRow(context.Background(), "SELECT id, movie_show_id, seat_id, ticket_status, price FROM tickets WHERE id = $1", id).
 			Scan(&t.ID, &t.MovieShowID, &t.SeatID, &t.Status, &t.Price)
 
 		if IsError(w, err) {
@@ -132,18 +149,22 @@ func GetTicketByID(db *pgxpool.Pool) http.HandlerFunc {
 // @Router /tickets/{id} [put]
 func UpdateTicket(db *pgxpool.Pool) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		id := r.PathValue("id")
+		id, ok := ParseUUIDFromPath(w, r.PathValue("id"))
+		if !ok {
+			return
+		}
+
 		var t Ticket
 		if !DecodeJSONBody(w, r, &t) {
 			return
 		}
-		t.ID = id
+		t.ID = id.String()
 
 		if !validateTicketData(w, t) {
 			return
 		}
 
-		res, err := db.Exec(context.Background(), "UPDATE tickets SET movie_show_id=$1, seat_id=$2, ticket_status_id=$3, price=$4 WHERE id=$5",
+		res, err := db.Exec(context.Background(), "UPDATE tickets SET movie_show_id=$1, seat_id=$2, ticket_status=$3, price=$4 WHERE id=$5",
 			t.MovieShowID, t.SeatID, t.Status, t.Price, t.ID)
 		if IsError(w, err) {
 			return
@@ -165,7 +186,11 @@ func UpdateTicket(db *pgxpool.Pool) http.HandlerFunc {
 // @Router /tickets/{id} [delete]
 func DeleteTicket(db *pgxpool.Pool) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		id := r.PathValue("id")
+		id, ok := ParseUUIDFromPath(w, r.PathValue("id"))
+		if !ok {
+			return
+		}
+
 		res, err := db.Exec(context.Background(), "DELETE FROM tickets WHERE id = $1", id)
 		if IsError(w, err) {
 			return
@@ -178,40 +203,43 @@ func DeleteTicket(db *pgxpool.Pool) http.HandlerFunc {
 	}
 }
 
-// @Summary Получить билеты пользователя
-// @Tags tickets
-// @Produce json
-// @Param user_id path string true "ID пользователя"
-// @Success 200 {array} Ticket
-// @Failure 500 {object} ErrorResponse "Ошибка"
-// @Router /tickets/user/{user_id} [get]
-func GetTicketsByUserID(db *pgxpool.Pool) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		userID := r.PathValue("user_id")
+// // @Summary Получить билеты пользователя
+// // @Tags tickets
+// // @Produce json
+// // @Param user_id path string true "ID пользователя"
+// // @Success 200 {array} Ticket
+// // @Failure 500 {object} ErrorResponse "Ошибка"
+// // @Router /tickets/user/{user_id} [get]
+// func GetTicketsByUserID(db *pgxpool.Pool) http.HandlerFunc {
+// 	return func(w http.ResponseWriter, r *http.Request) {
+// 		userID, ok := ParseUUIDFromPath(w, r.PathValue("user_id"))
+// 		if !ok {
+// 			return
+// 		}
 
-		rows, err := db.Query(context.Background(), `
-			SELECT id, movie_show_id, seat_id, ticket_status_id, price
-			FROM tickets
-			WHERE user_id = $1`, userID)
-		if IsError(w, err) {
-			return
-		}
-		defer rows.Close()
+// 		rows, err := db.Query(context.Background(), `
+// 			SELECT id, movie_show_id, seat_id, ticket_status, price
+// 			FROM tickets
+// 			WHERE user_id = $1`, userID)
+// 		if IsError(w, err) {
+// 			return
+// 		}
+// 		defer rows.Close()
 
-		var tickets []Ticket
-		for rows.Next() {
-			var t Ticket
-			if err := rows.Scan(&t.ID, &t.MovieShowID, &t.SeatID, &t.Status, &t.Price); err != nil {
-				http.Error(w, "Ошибка при сканировании", http.StatusInternalServerError)
-				return
-			}
-			tickets = append(tickets, t)
-		}
+// 		var tickets []Ticket
+// 		for rows.Next() {
+// 			var t Ticket
+// 			if err := rows.Scan(&t.ID, &t.MovieShowID, &t.SeatID, &t.Status, &t.Price); err != nil {
+// 				http.Error(w, "Ошибка при сканировании", http.StatusInternalServerError)
+// 				return
+// 			}
+// 			tickets = append(tickets, t)
+// 		}
 
-		if len(tickets) == 0 {
-			http.Error(w, "Билеты не найдены", http.StatusNotFound)
-		}
+// 		if len(tickets) == 0 {
+// 			http.Error(w, "Билеты не найдены", http.StatusNotFound)
+// 		}
 
-		json.NewEncoder(w).Encode(tickets)
-	}
-}
+// 		json.NewEncoder(w).Encode(tickets)
+// 	}
+// }
